requirements.txt
glfw == 2.0.0
PyOpenGL == 3.1.5